ensure that the plugin is not run on dependencies
Right now in rebar plugins are expected to run on all apps in the
repo. That includes dependencies. There is no really easy way to get
around that. To avoid that the system now looks for a version of
"semver". In those that contain this tag we replace the version with a
good semver version.